+++ offer-manager-domain/src/main/groovy/life/qbic/business/offers/update/UpdateOffer.groovy
@@ -1,14 +1,13 @@
 package life.qbic.business.offers.update
 
 import life.qbic.business.Constants
-import life.qbic.business.exceptions.DatabaseQueryException
 import life.qbic.business.logging.Logger
 import life.qbic.business.logging.Logging
 import life.qbic.business.offers.Converter
 import life.qbic.business.offers.Offer
 import life.qbic.business.offers.create.CreateOfferDataSource
-import life.qbic.business.offers.create.CreateOfferOutput
 import life.qbic.business.offers.identifier.OfferId
+import life.qbic.business.offers.identifier.OfferIdDtoMapper
 
 /**
  * <h1>SRS - 4.2.2 Update Offer</h1>
@@ -83,10 +82,8 @@ class UpdateOffer{
     private void fetchAllAvailableVersions() {
         def versions = dataSource.fetchAllVersionsForOfferId(Converter.convertIdToDTO(offerToUpdate
                 .identifier))
-        offerToUpdate.addAllAvailableVersions(
-                versions.stream()
-                        .map(version -> Converter.buildOfferId(version))
-                        .collect())
+        List<OfferId> offerIds = versions.stream().map(OfferIdDtoMapper.DTO_TO_OFFER_ID).collect()
+        offerToUpdate.addAllAvailableIdentifiers(offerIds)
     }
 
     private static Offer createBusinessOffer(life.qbic.datamodel.dtos.business.Offer offer){
@@ -108,36 +105,8 @@ class UpdateOffer{
         return createBusinessOffer(offerDTO)
     }
 
-    private OfferId createNewVersionTag(){
-
-        //search for all ids in the database
-        List<OfferId> allVersionIds = dataSource
-                .fetchAllVersionsForOfferId(Converter.convertIdToDTO(offerToUpdate.identifier))
-                .stream().map(offerId -> Converter.buildOfferId(offerId))
-                .collect()
-
-        offerToUpdate.addAllAvailableVersions(allVersionIds)
-
-        OfferId convertedId = offerToUpdate.getLatestVersion()
-        convertedId.increaseVersion()
-
-        return convertedId
-    }
-
     private static boolean theOfferHasChanged(Offer oldOffer, Offer newOffer) {
         return oldOffer.checksum() != newOffer.checksum()
-    }
-
-    private static Offer buildOffer(life.qbic.datamodel.dtos.business.Offer offer, OfferId identifier){
-        return new Offer.Builder(
-                        offer.customer,
-                        offer.projectManager,
-                        offer.projectTitle,
-                        offer.projectDescription,
-                        offer.items,
-                        offer.selectedCustomerAffiliation)
-                        .identifier(identifier)
-                        .build()
     }
 
 }