+++ CHANGELOG.rst
@@ -13,12 +13,13 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Fixed**
 
+* Offer identifiers are now all in the format `O_project_rand_1`
+
 * ZIP code and city of the offer recipient are in the same line now
 
 **Dependencies**
 
 **Deprecated**
-
 
 1.2.1 (2021-10-27)
 ------------------